+++ build.sh
@@ -18,11 +18,11 @@ check_status () {
   return
 }
 
-printf "\n\n [*]  Ruuning 'dart pub update' to check packages are current...\n\n"
+printf "\n\n [*]  Running 'dart pub update' to check packages are current...\n\n"
 dart pub update
-printf "\n\n [*]  Ruuning 'dart foramt' to check source code files...\n\n"
+printf "\n\n [*]  Running 'dart format' to check source code files...\n\n"
 dart format --output=none --set-exit-if-changed .
-printf "\n\n [*]  Ruuning 'dart analyse' to check source code files...\n\n"
+printf "\n\n [*]  Running 'dart analyse' to check source code files...\n\n"
 dart analyze
 printf "\n\n [*]  Building '%s'...\n\n" "$DARTAPP"
 if [[ ! -d ./build ]]; then